Improvements
This updates the cherri binary to the latest build. This makes the
editor support tabs. This adds a settings window. This adds indication
of compilation status. This adds a banner to the editor when the
compiler is unable to generate a Shortcut. This fixes and improves
adding inline messages by improvements to parsing the compiler output.

The toolbar now no longer has a settings popover. The toolbar now
contains the documentation link as well as a build and now run button.
This is to make running the compiled Shortcut in the Shortcuts app
optional between build and run.